requirements.txt
gTTS==2.5.2
matplotlib==3.9.1.post1
Pillow==10.4.0
playsound==1.3.0
pyttsx3==2.90
Requests==2.32.3
tinydb==4.8.0
